fixed ifft bug

+++ trunk/inv_pfb/visualizer.py
@@ -4,8 +4,8 @@ import scipy.signal
 if len(sys.argv) != 3:
 	print "please specify [tone file] and [output file]"
 	sys.exit(1)
-tone = np.fromfile(sys.argv[1])
-output = np.fromfile(sys.argv[2])
+tone = np.fromfile(sys.argv[1],dtype=np.float32)
+output = np.fromfile(sys.argv[2],dtype=np.float32)
 xc = scipy.signal.correlate(tone,output)
 figure(1)
 subplot(211)
